feat: SOF-1130 change HTML folder using Graphics Setups
